requirements.txt
dalybms==0.5.0
pyserial==3.5
GPIOSimulator==0.1